split annotationlib into two
